requirements.txt
flask==2.0.1
python-dotenv==0.19.1
google-api-python-client==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6